refactor: login, register controller 하나로 통합
