--- app/src/main/java/com/techcourse/controller/RegisterViewController.java
@@ -1,16 +1,0 @@
-package com.techcourse.controller;
-
-import com.interface21.context.stereotype.Controller;
-import com.interface21.web.bind.annotation.RequestMapping;
-import com.interface21.web.bind.annotation.RequestMethod;
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.view.JspView;
-
-@Controller
-public class RegisterViewController {
-
-    @RequestMapping(value = "/register/view", method = RequestMethod.GET)
-    public ModelAndView showRegisterView() {
-        return new ModelAndView(new JspView("/register.jsp"));
-    }
-}
